Isolate MCP creation preparation
