+++ apps/server/src/mcp.ts
@@ -5,13 +5,18 @@
  * module only validates and presents their results through MCP.
  */
 
-import { createHash } from "node:crypto";
+import {
+	BRIEF,
+	isRepository,
+	MAX_REQUEST_BYTES,
+	MAX_TITLE_LENGTH,
+	prepare,
+	REPOSITORY_PATH_PATTERN,
+} from "./mcp/create";
 
-import { lookup, parse, serialize, ulid, validate } from "@chopin/dialect";
+import type { Brief, CreateDocumentInput } from "./mcp/create";
 
-import * as room from "./plan/room";
-
-import type { Nodes, Parent, Root } from "mdast";
+export type { Brief, CreateDocumentInput, CreationOrigin } from "./mcp/create";
 
 export type DocumentSummary = {
 	id: string;
@@ -27,28 +32,6 @@ export type Document = DocumentSummary & {
 export type DocumentReader<Caller> = {
 	list(caller: Caller, repository: string): Promise<DocumentSummary[]>;
 	read(caller: Caller, id: string): Promise<Document | undefined>;
-};
-
-export type Brief = {
-	goal: string;
-	constraints: string[];
-	settledDecisions: string[];
-	openQuestions: string[];
-	repositoryFindings: string[];
-};
-
-export type CreationOrigin = {
-	idempotencyKey: string;
-	fingerprint: string;
-	repository: string;
-	baseBranch: string;
-	baseCommit: string;
-	title: string;
-};
-
-export type CreateDocumentInput = CreationOrigin & {
-	brief: Brief;
-	plan: string;
 };
 
 export type CreatedDocument = Document & { url: string };
@@ -79,15 +62,7 @@ type Tool = {
 	outputSchema: Record<string, unknown>;
 };
 
-const OWNER_PATTERN = "[A-Za-z0-9](?:[A-Za-z0-9]|-(?=[A-Za-z0-9])){0,38}";
-const REPOSITORY_PATTERN = "(?!\\.\\.?$)[A-Za-z0-9._-]{1,100}";
-const REPOSITORY_PATH_PATTERN = `^${OWNER_PATTERN}/${REPOSITORY_PATTERN}$`;
-const OWNER = new RegExp(`^${OWNER_PATTERN}$`);
-const REPOSITORY = new RegExp(`^${REPOSITORY_PATTERN}$`);
 const MAX_DOCUMENT_ID_LENGTH = 128;
-const MAX_TITLE_LENGTH = 120;
-/** Read-only JSON-RPC calls need arguments, not document-sized request bodies. */
-const MAX_REQUEST_BYTES = 64 * 1_024;
 
 const DOCUMENT = {
 	type: "object",
@@ -96,19 +71,6 @@ const DOCUMENT = {
 		title: { type: "string" },
 	},
 	required: ["id", "title"],
-	additionalProperties: false,
-};
-
-const BRIEF = {
-	type: "object",
-	properties: {
-		goal: { type: "string", minLength: 1 },
-		constraints: { type: "array", items: { type: "string" } },
-		settledDecisions: { type: "array", items: { type: "string" } },
-		openQuestions: { type: "array", items: { type: "string" } },
-		repositoryFindings: { type: "array", items: { type: "string" } },
-	},
-	required: ["goal", "constraints", "settledDecisions", "openQuestions", "repositoryFindings"],
 	additionalProperties: false,
 };
 
@@ -252,122 +214,10 @@ function toolCall(
 	return arguments_ ? { name: value.name, arguments: arguments_ } : undefined;
 }
 
-function isRepository(value: unknown): value is string {
-	if (typeof value !== "string") return false;
-	let parts = value.split("/");
-	return parts.length === 2 && OWNER.test(parts[0]!) && REPOSITORY.test(parts[1]!);
-}
-
 function isId(value: unknown): value is string {
 	return typeof value === "string"
 		&& Array.from(value).length <= MAX_DOCUMENT_ID_LENGTH
 		&& value.trim().length > 0;
-}
-
-function strings(value: unknown): string[] | undefined {
-	return Array.isArray(value) && value.every(item => typeof item === "string") ? value : undefined;
-}
-
-function nonblank(value: unknown, maximum: number): string | undefined {
-	return typeof value === "string"
-			&& Array.from(value).length <= maximum
-			&& value.trim().length > 0
-		? value
-		: undefined;
-}
-
-type CreateArguments = Omit<CreateDocumentInput, "fingerprint">;
-
-function createArguments(value: Record<string, unknown>): CreateArguments | undefined {
-	let expected = [
-		"idempotencyKey",
-		"repository",
-		"baseBranch",
-		"baseCommit",
-		"title",
-		"brief",
-		"plan",
-	];
-	if (
-		Object.keys(value).length !== expected.length
-		|| expected.some(key => !Object.hasOwn(value, key))
-	) return undefined;
-	let brief = record(value.brief);
-	let goal = nonblank(brief?.goal, 4_096);
-	let constraints = strings(brief?.constraints);
-	let settledDecisions = strings(brief?.settledDecisions);
-	let openQuestions = strings(brief?.openQuestions);
-	let repositoryFindings = strings(brief?.repositoryFindings);
-	let idempotencyKey = nonblank(value.idempotencyKey, 128);
-	let baseBranch = nonblank(value.baseBranch, 255);
-	let baseCommit = nonblank(value.baseCommit, 64);
-	let title = nonblank(value.title, MAX_TITLE_LENGTH);
-	let plan = nonblank(value.plan, MAX_REQUEST_BYTES);
-	if (
-		!idempotencyKey
-		|| !isRepository(value.repository)
-		|| !baseBranch
-		|| !baseCommit
-		|| !title
-		|| !goal
-		|| !constraints
-		|| !settledDecisions
-		|| !openQuestions
-		|| !repositoryFindings
-		|| !plan
-	) return undefined;
-	return {
-		idempotencyKey,
-		repository: value.repository,
-		baseBranch,
-		baseCommit,
-		title,
-		brief: { goal, constraints, settledDecisions, openQuestions, repositoryFindings },
-		plan,
-	};
-}
-
-function identify(node: Nodes | Root): void {
-	if (node.type === "mdxJsxFlowElement" || node.type === "mdxJsxTextElement") {
-		let component = lookup(node.name);
-		if (
-			component?.attributes.id?.required
-			&& !node.attributes.some(attribute =>
-				attribute.type === "mdxJsxAttribute" && attribute.name === "id"
-			)
-		) node.attributes.unshift({ type: "mdxJsxAttribute", name: "id", value: ulid() });
-	}
-	if ("children" in node) {
-		for (let child of (node as Parent).children) identify(child);
-	}
-}
-
-function canonical(source: string):
-	| { source: string }
-	| { issues: Array<{ code: string; message: string; path: string; offset?: number }> }
-{
-	let tree: Root;
-	try {
-		tree = parse(source);
-	} catch (err) {
-		return {
-			issues: [{
-				code: "parse",
-				message: err instanceof Error ? err.message : String(err),
-				path: "root",
-			}],
-		};
-	}
-	identify(tree);
-	let result = validate(tree);
-	if (!result.ok) return { issues: result.issues };
-	let output = serialize(tree);
-	room.validate(output);
-	return { source: output };
-}
-
-function fingerprint(input: CreateArguments): string {
-	return createHash("sha256").update(JSON.stringify(input)).digest("hex");
 }
 
 function isJsonRpcId(value: unknown): value is string | number | null {
@@ -501,19 +351,14 @@ export function handler<Caller>(
 					return document ? respond(text(document)) : respond({ content: [], isError: true });
 				}
 				if (tool.name === "create_document") {
-					let input = createArguments(tool.arguments);
-					if (!input) {
+					let prepared = prepare(tool.arguments);
+					if (!prepared) {
 						return notification
 							? undefined
 							: error(call.id, -32602, "create_document requires a valid draft");
 					}
-					let prepared = canonical(input.plan);
 					if ("issues" in prepared) return respond(text({ issues: prepared.issues }, true));
-					let outcome = await options.create.create(caller, {
-						...input,
-						plan: prepared.source,
-						fingerprint: fingerprint(input),
-					});
+					let outcome = await options.create.create(caller, prepared.input);
 					if (outcome.kind === "created" || outcome.kind === "replayed") {
 						return respond(text(outcome.document));
 					}
